move optimized parsing to statusandheaders
